Fix .t to handle Perl devel changes
Changes to core Perl during the v5.21 development series have broken
both test files in this module when run under a locale whose decimal
radix character is not a dot.  The solution is for the tests to just
make sure the locale is set to "C" before proceeding with the testing.

See https://rt.perl.org/Ticket/Display.html?id=123169

The changes to core Perl were necessary to get localeconv() to work
consistently.  Previously, it was somewhat haphazard whether it took
effect outside of 'use locale' or not.  Hence, just uncommenting the
lines for trying a Swedish locale in your files might not have worked
prior to the core Perl changes.